+++ sale_line_price_properties_based/__init__.py
@@ -19,4 +19,4 @@
 #
 ##############################################################################
 from . import product
-from . import sale
+from . import sale_order_line
