+++ sale_line_price_properties_based/sale_order_view.xml
