+++ sale_line_quantity_properties_based/__init__.py
@@ -18,5 +18,5 @@
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
 ##############################################################################
-from . import sale
+from . import sale_order_line
 from . import product
