+++ sale_properties_dynamic_fields/mrp_property_group_view.xml
