+++ sale_properties_easy_creation/mrp_property_demo.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+    <data noupdate="1">
+        <record id="length_5" model="mrp.property">
+            <field name="name">Length 5</field>
+            <field name="group_id" ref="length_group"></field>
+            <field name="value">5</field>
+        </record>
+        <record id="width_1" model="mrp.property">
+            <field name="name">Width 1</field>
+            <field name="group_id" ref="width_group"></field>
+            <field name="value">1</field>
+        </record>
+        <record id="painting" model="mrp.property">
+            <field name="name">Painting</field>
+            <field name="group_id" ref="painting_group"></field>
+            <field name="value">True</field>
+        </record>
+    </data>
+</openerp>
